ai-edi-orchestration/requirements.txt
fastapi
uvicorn
numpy
psycopg2-binary
python-dotenv
asyncpg
python-jose
requests
python-json-logger